+++ NEWS.md
@@ -1,3 +1,6 @@
+# `tidysdm` dev
+* resolve conflicting coords names in `thin_by_cell()`
+
 # `tidysdm` 1.0.4
 * improve documentation of `calib_class_threshold` and implement 
   `collect_calib_thresh`